Initial Commit
I have been playing around with creating a program to generate a school
calendar for year-round homeschooling. I've built a handful of functions
that will be needed, but I do not have a clear plan of action in place
for how this will ultimately come together.

Before going any further with the project, I decided it would be wise to
go ahead and get a repository going for it.

I've created a simple README and included the beginnings of the program
in this commit.

modified:   README.md
new file:   date.c
new file:   input.c