Remove all code except headers from input.c
input.c currently only has one function written into it. This function
relies on the scanf() and printf() functions to get from and put to the
terminal. I want to rewrite this function and determined it would be
best while the file is still small to just start from scratch.