Clean up data types in date.h and date.changes
Used typedef to declare data type of Date instead of having struct date
everywhere in the code. I removed the date_range struct and weekdays
enum as they were not actually being used anywhere in the code. I also
removed the months enum and replaced the instances where the code was
using it with a local const int FEBRUARY = 2.